machineTest/problem2/restApi/requirements.txt
Django==4.1.6
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
mongoengine==0.26.0
Pillow==9.4.0
PyJWT==1.7.1
pymongo==3.12.1
pytz==2022.7.1